move sensor time into separate field
it makes more sense to keep the ROS time and the sensor timestamp separate, e.g. for debugging purposes. E.g. for mech. scanning sonars the start time of acquiring a scan and the time the message is sent differ significantly, which may lead to problems e.g. when re-playing a rosbag.